backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary 
alembic 
python-dotenv 
pydantic 
passlib[bcrypt] 
bcrypt==4.0.1
python-jose
email-validator
python-multipart
requests
google-auth
pytest
httpx

